% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2latex.R
\name{format_subscript}
\alias{format_subscript}
\title{Adding index in LaTeX}
\usage{
format_subscript(term)
}
\arguments{
\item{text}{}
}
\value{
text
}
\description{
Adding index in LaTeX
}
\examples{
format_subscript('X17')
}
