% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2latex.R
\name{model_to_latex.glm}
\alias{model_to_latex.glm}
\title{Generalized linear model method to LaTeX}
\usage{
model_to_latex.glm(model)
}
\arguments{
\item{model}{}
}
\value{
string
}
\description{
Generalized linear model method to LaTeX
}
\examples{
# Generalized linear model method with family-specific checks
}
