% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2latex.R
\name{model_to_latex.lm}
\alias{model_to_latex.lm}
\title{Linear model method to LaTeX}
\usage{
model_to_latex.lm(model)
}
\arguments{
\item{model}{}
}
\value{
string
}
\description{
Linear model method to LaTeX
}
