% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2latex.R
\name{sentence_case}
\alias{sentence_case}
\title{String to Sentence Case}
\usage{
sentence_case(text)
}
\arguments{
\item{text}{}
}
\value{
text
}
\description{
String to Sentence Case
}
\examples{
sentence_case('asdasdasd')
}
